Great starting progress
Created the bones of the project by using Router with Routes and each Route path matches a page created. The links to these pages are found in the AppNavbar.js component. I imported and ever so slightly tweaked last week's project to be its own page for this project named DinosaurDatabase.js. All pages and the error page are working as intended.