Updated pictures/database
Properly got the database to display the dinosaur's picture (updated the if statement and the props in the onSubmit function in NewDinosaurForm.js. Added many more pictures to choose from. Next goal is to work on the dinosaur battles page (hint: dropdown lists for two dinosaurs, using random number generators for HP and attacks).