Updated DinosaurBattles
Put lots of work into the beginnings of the DinosaurBattles.js page. The database is fetched, but much less info is displayed about the dinos. This dinosaur is being returned with an additional prop of attack, which will be used when the battle system begins to be implemented. For now, I need to work on updating the state and displaying what I want after the dinosaur is chosen.