More DinosaurBattles progress
Updated the API to contain another resource called DinosaurBattles. Created a DinosaurBattlesAPI.js that will handle the DinosaurBattles data. Worked on the readyForBattle function within MinimalDinosaursList.js and created a function createBattleDinosaur that makes an asynchronous API call to the DinosaurBattlesAPI. Created comments to reflect next steps needed for this project.