Currently readyForBattle not working
State is not being read on current line 51 in MinimalDinosaursList.js. There is an issue relating to state. I am trying to check to see if the array of battlingDinosaurs contains less than 2 objects. Then and only then would I want to allow an update to the dinosaurBattlesAPI. Need to find a way to fix this. I have added the random number generator for health and power, and I have added new API calls. Currently added a constructor to include the state, instead of how it was before as just an object state outside of a constructor.