Fixed issues with state!
State was not able to be checked in the readyForBattle function within MinimalDinosaursList.js because the function was being passed down to the child component MinimalDinosaur. I moved the state check to the createBattleDinosaur function in MinimalDinosaurList.js. Current issue is the state not properly reflecting that a battlingDinosaur has been added to the API, thus the state list being one smaller than it should. Also the battleReadyDino object is being passed into the API as one whole object, with the pre-set schema getting seemingly random values and then that object added as an additional value object with all the correct values.