Fixed API issue and fixed state issue
dinosaurBattlesAPI.post object argument was changed to remove the {} around battleReadyDinosaur and is now properly accepting all the values.  Changed state to be updated off of the fetch response promise, now assigned to the const battlingDinosaurs. Added the if statements and do/while loop to begin and end battles between the dinosaurs. Have no implemented the API call to delete contents of the battlingDinosaursAPI. Also ideally the battle will not happen in just console logs.